- newer gcc warned me about errors in our strncat usage - this function is very   unintuitive as it writes up to n+1 bytes, so you have to do:   	strncat(buf, X, sizeof(buf)-1); - fixed a bunch of other warnings in release mode
[r25996]

+++ ir/ir/irprog.c
@@ -322,7 +322,7 @@ void add_irp_opcode(ir_op *opcode) {
 	assert(irp);
 	len  = ARR_LEN(irp->opcodes);
 	code = opcode->code;
-	if (code >= len) {
+	if ((int) code >= len) {
 		ARR_RESIZE(ir_op*, irp->opcodes, code+1);
 		memset(&irp->opcodes[len], 0, (code-len+1) * sizeof(irp->opcodes[0]));
 	}
